Add a method for graphviz-ing an interval tree
Numerous algorithms on the IntervalTree have complex and nuanced
pathological cases that can be really difficult to debug - having the
ability to *visualize* the tree makes it singificantly easier to debug
these issues. This commit adds a method to IntervalTree to print the
graphviz `dot` representation of a tree, which I've been printing out,
copying, and piping to `dot` during development to help debug edge
cases.