AVL-balance IntervalTree on insertion
Implement the AVL self-balancing binary tree algorithm for IntervalTree
on insertion, but not removal (yet). Concretely:

- We add a new `height` field to `Node` that tracks the overall height
  of that node (aka, one more than the max of its children's heights)
- On insertion, while walking down the tree we save pointers to the
  nodes we've visited on our way down, and then on our way back up we
  rebalance the node using the AVL rebalancing algorithm (left-right,
  right-right, left-left, or right-right).
- On removal, we update the heights of the nodes, but otherwise don't
  rebalance yet

Rebalancing on insert uncovered a bug in removal where in certain
cases (concretely, a node that shares a start interval with an indirect
descendant but not its immediate child) we could end up with either
duplicate intervals or an unsorted tree. This can happen if, while
subtracting an interval from a node, we *reduce* the upper bound of that
node (either due to splitting it or just shrinking the top) such that it
is no longer greater than its ancestors, even though it's on the right
side. Now, rather than attempting to reuse nodes after removing from
them, if we *ever* change the interval of a node in any way we just
split that node off (by returning `true` from walk_tree) and add its
children to the queue to re-insert.

Finally, since it's completely unused, rather than maintaining
remove_random_leaf and updating it to rebalance after removal, this
removes it entirely.