Move evbtree to dataflow, and rename to reader_map
The `evbtree` data structure, which started out as a not necessarily
noria-specific, general-purpose data structure (wrapping a hash map!)
has been growing less general-purpose and more noria-specific as time
has passed. The addition of an interval tree didn't *quite* tip the
scales for me, but soon I'm going to be doing some work to make the
backing data structure dynamic (either a HashMap or a BTreeMap), after
which it feels like it's going to be *really* not general-purpose enough
to live as its own top-level crate - not to mention that calling it an
evbtree will be a total falsehood! This commit moves the evbtree crate
inside of noria/server/dataflow, and renames it to reader_map to be more
explicit about what it's responsible for.