Make reader_map's backing store dynamic
Introduce a new enum, called Data, for the data structure backing a
reader_map (previously known as evbtree), and wrapping either a HashMap
or the combination of a BTreeMap and an IntervalTree. All operations
that both data structures can support (insert, get, etc.) are supported
via pattern matching on the enum. Operations that are only supported by
the BTreeMap (range, remove_range, etc.) panic when attempted on a
reader_map backed by a HashMap - the expectation here is that any
calling code will have set up the appropriate index type ahead of time,
so this should be a program invariant.

To indicate to the constructor which index type we want to use, this
commit also introduces a new IndexType enum to noria::internal, with
HashMap and BTreeMap members.

Unfortunately, since this introduces a dependency on noria in
reader_map, we can no longer check clippy on reader_map in CI, since
noria doesn't pass clippy - eventually we should fix that, but for now
this just removes it.