Make KeyedState's backing store dynamic
Similarly to what we did for reader_map, make the KeyedState enum (which
already had different members for maps with keys of different lengths)
support being backed by either a BTreeMap (wrapped in a PartialMap to
support the IntervalTree of materialized ranges) or a HashMap. To
indicate this during construction, this introduces a new Index struct
containing an IndexType and a Vec<usize> of column indices, which is now
supplied as the first parameter to State::add_key (PersistentState
ignores the IndexType, since rocksdb can always support ranges). passing
around that Index struct has been threaded up through the domain,
through the various payload enums we use to tell domains about their
nodes' keys, and into the materialization planner, where it's been left
constantly specified as IndexType::BTreeMap along with a TODO.