Teach readers about their index type
Thread information about the index type for a reader, determined during
MIR by looking at the binary operators used on parameters, all the way
through MirNodeInner::Leaf and into the construction of the reader
itself, where instead of storing a Vec<usize> for the key columns of the
reader we store a full-fledged Index. As of this commit nothing is
actually using this, though.

Change-Id: I543111f781f482c811b4f1389be9acc447e2120d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/730
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>