Refactor noria-mysql backend into new top-level noria-client crate
This refactor moves the noria-mysql/backend module (and associated
modules) into its own noria-client crate. This is the first step in
creating a noria client independent from existing mysql clients.
The current vision for this crate is a client library that can be easily
wrapped in other languages and packaged for use in a variety of
applications.

All existing functionality is preserved, and the noria-mysql crate now just
consists of main.rs to run the noria-mysql adapter.

All integration tests associated with the library have moved from noria-mysql
to noria-client as well.