Always use hashmaps, and disable range query tests
We'd like to temporarily stop work on index selection for now, at least
the parts of it that support range queries, and bring noria back to a
the performance level we had with hashmaps. This commit marks all tests
that depend on range queries in some way (including ilike) as ignored,
and hardcodes hashmaps instead of btreemaps as the default index type we
use for all replay obligation indices in materialization.