Add simplistic metrics framework for noria-server
The `metrics` crate seems to have a pretty nice framework for recording
metrics. We implement a `metrics::Recorder` (`NoriaMetricsRecorder`) based
on an operation log and a mutex of hashmaps; metrics changes get
serialized to the operation log (a lock-free data structure from the
`crossbeam` crate), and are applied (via 'collapse()') to the hashmaps
whenever the operation log is full, or the metrics are read.

The `/metrics_dump` route in noria-server gets the current metrics state
and returns it as a JSON blob, using the `MetricsDump` struct. As a
small test, making external requests to Noria increments the
`server.external_requests` counter.

clubhouse: [ch292]