Add a generator for supported SQL queries
Add a first crack at a deterministic, exhaustive, parametric generator
for SQL queries, and associated DDL. The idea is that we can plug these
queries (or a subset of them, there are quite a few!) into a
benchmarking harness to determine the specific performance costs of
specific combinations of operators in Noria. Also, at some point in the
future, we could use these generated queries to generate sqllogictest
scripts for comparing Noria's behavior against that of something like
MySQL.

The general approach here:

- We have a QueryOperation enum, which enumerates, in some sense, the
  individual "operations" that can be performed as part of a SQL query
- Each QueryOperation knows how to add itself to a SQL query
- To support that, there's a "GeneratorState" struct, to which mutable
  references get passed around, which knows how to summon up new tables
  and columns for use in queries
- We can then calculate all permutations of all the possible
  QueryOperations (up to a certain depth), and use those to generate
  queries.

Not all operations are supported yet - the current version has
aggregates, distinct, constant and parametrized filters, literal
projections, and joins, but in the future I'd also like to add
arithmetic projections, topk, union, order by, and ilike. Also, as we
add support for additional query semantics, adding those operators here
should be relatively straightforward.

For use in manual testing and verification I've attached a small CLI to
this, with the ability to run and generate queries given a maximum
number of operations. With the current set of supported operators this
generates 7806 queries.