Strip out non-public serde exports from data_serde
The serde Serialize and Deserialize impls of DataType were originally
generated via a manually-edited expansion of the derive macros for both
of those traits - this resulted in some unidiomatic code (including
identifiers prefixed by two underscores for hygiene) but more notably
resulted in code that stopped working after an upgrade of a minor
version of serde, since it depended on a module (serde::export) that was
not part of the public interface of serde, and has since been removed.
This commit *slightly* rewrites those implementations to remove some of
the double-underscore-prefixed identifiers, and also no longer depend on
serde::export at all.