Clarify the commit lint error for message bodies
I've noticed some people confused by the current wording of the error
given by the commit linter for commit message bodies - specifically, we
never actually say what a "body" is or how to give a commit a body. This
adds some extra description to the error explicitly saying what should
be in a commit message body, and how to format a commit such that it has
a body.