Support debezium to base table data flow node timestamp propagation
This refactors the table service to support an additional request
type: Timestamp. This allows constructing a packet data which
contains a timestamp, and sends it to the base table node. The domain
drops the packet upon handling. A future change will handle the
packet at data flow nodes and store relevant timestamp information.

This change makes several small changes to applications that call
directly into the table service due to the refactor.