Add functions to generate seed data for queries
Add support to query_generator to generate seed data for queries, based
on the schema of the tables that have been created for those queries. To
ensure that constant-valued filters get at least some results, there's a
new `expected_values` map on the GeneratorState, which QueryOperations
can populate when emitting query ops, and which values from are used
half the time when generating data.