Propagate base table timestamps along dataflow graph to readers.
This change adds timestamp packet handling at each node in the dataflow
graph. The notable cases are (1) the base table node which handles
proxying the timestamp and ack'ing a request from debezium, (2)
the reader node which gets and sets the timestamp values from the
reader map, and (3) every other node which sends the timestamp
to all it's children. At each node the `min` timestamp function is
used to calculate the timestamp to propagte to their children.