Decouple backend struct's core functions from MysqlShim implementation
For context, the `backend` struct in `noria-client` was originally written
exclusively to work with `msql-srv`. Therefore, to communicate
with noria and/or an external DB, `backend` basically just implemented three
functions as a part of the `MysqlShim` trait: `on_prepare`, `on_execute`,
and `on_query`.

The goal is to not necessarily use `msql-srv` anymore when using this client.
In this change, the "core functionality" of these three functions is
ripped out into three new functions: `prepare`, `execute`, and `query`.
These three new functions belong to the backend struct, and take care of
parsing and carrying out queries now.

`prepare` returns a new enum `PrepareResult` wrapped in a `Result`.
`execute` and `query` return a new enum `QueryResult` wrapped in a `Result`.

The three original MysqlShim trait functions now call one of the three new
functions under the hood and basically are just wrappers.

These three new functions provide `backend` with an API that doesn't rely on
`msql-srv`. All future clients, such as a javascript client,
are expected to wrap and flow through this core API.

The MysqlShim trait implementation should likely be moved out of
`noria-client` and implemented seperately in `noria-mysql` in the future.

Also created a small example in noria-client querying directly with
the backend struct without `msql-srv` running.