Allow any quoted string in group_concat separator
The parser for the separator of a group_concat previously only allowed
alphanumeric characters, rather than just reusing the parser for quoted
strings - this meant that attempting to use a non-alphanumeric separator
would fail to parse. This updates it to instead just use the existing
single- or double-quoted string parser.