Instrument upquery time and materialization overhead (and more)
Various counter!() and gauge!() calls log the amount of time spent doing various
upquery-related tasks. This is done with a high degree of granularity
(per-domain, per-shard, and per-tag), which should provide the future metrics
harness with all of the information it needs.

NOTE: In future, we probably want to do some form of aggregation on the metrics
before feeding them into something like Prometheus, because having that degree
of granularity in actual time-series databases is considered a bad idea!

In addition, we log materialization overhead by adding a gauge for every piece
of state in a domain's `self.state`, as well as the total count. We also hook
into the eviction worker and add gauges for the total amount of *partial* state,
and a counter for every time it does an eviction.