Qualify all generated column references
Since generated column names are only unique per-table rather than
globally, sometimes we generate queries with ambiguous column names in
the case of queries that include multiple tables due to joins. The
safest thing to do there is to just qualify all column references in
generated queries with the names of their tables.