Add start of a benchmark runner to query_generator
Add the beginnings of a benchmark runner mode to query_generator. This
is (currently) exposed as a subcommand taking a list of operations
to benchmark. It generates a list of permutations of those operations,
and for each permutation generates a query, sets up a new noria
instance, installs the recipe for the query, seeds data, runs the query,
and prints timing. Once we have support for metrics, this will be
extended to scrape those metrics and print those results as well.