Add a noria client RPC wrapper for metrics dumps
Add a `metrics_dump()` method to the noria client handle for accessing
the newly-added /metrics_dump RPC call. To support this, since we'll
want to return a deserialized version of the MetricsDump type, this
moves the definition of that type to noria/noria itself, under a new
top-level `metrics` module. Also (unfortunately, since it's not a good
use of HTTP semantics), since all noria RPCs are POSTs this also
required changing the method of the /metrics_dump endpoint to POST.