Collect and report metrics when benchmarking queries
Instrument the query benchmarking process with the collection and
reporting of a number of metrics, collected both via timing things
directly and via scraping the metrics_dump method on the noria handle
throughout the process.

Currently, this does two reads of the same key - one cold and one warm -
and uses those to collect memory usage and timing metrics for both cold
and warm caches of each run query. In between each query we call a new
`clear()` method on the metrics collector to reset everything back to 0.

Some queries - notably certain aggregate functions - currently fail when
benchmarked this way - to make that not break the entire benchmark we
explicitly handle any failures and package them up in a
QueryBenchmarkError struct, then continue going.