Add flag to update state sizes every packet
Add a new `aggressively_update_state_sizes` flag to the configuration
for a domain (threaded through from the overall noria server
configuration) to update the values for the memory used by in-memory
state after *every* packet is handled, rather than only when requested
by the eviction worker. This is intended for use in the benchmarking
harness (the one attached to the query generator), so that we can get
guaranteed-up-to-date metrics about materialization overhead and memory
usage without having to wait for the eviction worker to trigger an
update. Since the only intended usage is the benchmarking harness, and
that currently runs a noria server directly, this is only exposed via
noria_server::Builder and not via a command line flag or environment
variable, but that may want to change at some point