Update Taster to work with latest rust
This (relatively hefty, sorry) commit updates basically everything about
taster to work with the latest version of rustc, and the rust ecosystem
at large. The primary impetus for doing this at this point was that the
"afterparty" dependency, which provides webhook handling support,
depended on a pre-rust-2018 version of serde which used build-time code
generation rather than proc macros to generate (De)Serialize impls for
types - this transitively depended on a now-broken and unmaintained
version of the syntex_syntax crate, which was what was causing this to
not compile on a newer rustc (so much for backwards compatibility, I
suppose). An initial attempt to update that library to work with the
latest versions of all the various dependencies was stopped short after
a failed attempt to get Hyper's Service abstraction to work properly -
that abstraction is poorly documented, for one, and for another appears
to require usage of private traits to actually function - see
https://github.com/hyperium/hyper/issues/2148 (which has been
unresponded to and unresolved since september of last year). Given all
of that, and given that the actual webhook handling is a relatively
small part of Taster itself, I've decided to vendor the bits of
afterparty we care about (in webhook/events.rs) and just rip out Hyper
entirely, rewrite the HTTP request handling in rouille, which is
extremely simple, post-1.0 (so it won't have quite as many breaking
changes, hopefully) and much better documented.

Besides that, a lot of the rest of this commit consists of updating
things to use rust 2018 idioms, and generally work with newer versions
of all the dependencies.