Add deploy script for taster
Add the script that I'm currently using to deploy new versions of Taster
to AWS. This builds a release build of taster inside of a docker
image (so that the glibc etc. versions match up with what we're running
in aws), strips the binary, copies it up to the machine, and restarts
the service. Currently this can only be run by me (since only I have
auth on the instance) but is here for posterity and so that other people
given auth on the instance can run it