Don't treat same-table comparisons as joins
The checking in QueryGraph for whether a comparison in a WHERE was the
join key of an implicit join (what it refers to as a "comma-join") was a
little bit too aggressive - specifically, a comparison between two
columns is only the key of an  implicit join if those two columns are
from different tables - otherwise, it can be treated as  a regular old
global predicate. The main symptom of this was that any comparisons we
can't handle on join keys would cause the unimplemented() lower down to
hit, even though we can actually handle the query - in this case, this
was happening for a != on two columns of the same table.