Support benchmarks that output JSON
Add support to taster for benchmarks that output structured JSON
containing results of the benchmarks, specifically with an eye to the
output format of the query_generator benchmark, which outputs JSON
objects each containing a benchmark and several metrics.

To make this work, this breaks out the definition of the benchmark
results for a Benchmark config into a new OutputFormat enum, which has
members for both the old regex format and the new JSON output format.
There's an example of what the config format looks like in our own
taster.toml here, which has a definition for a benchmark run of the
query generator benchmark.

Since there's at least one test in taster now (for parsing the config)
this also adds a step to the buildkite config to run tests for taster