Use a unique column for query parameters
The MultipleParameters query operation is implemented by just delegating
to the SingleParameter query operation twice - this makes sense, except
that SingleParameter uses an *existing* column to filter on, rather than
making a fresh one, meaning that MultipleParameters ended up just
filtering on the same column twice. This changes the behavior of
SingleParameter to make a fresh column every time it's called.