Allow running with multiple sets of operations
Generally speaking, noria is buggy enough that running a full set of
permutations of query operations at the moment is basically a
non-starter, not to mention that it generates a *lot* of queries -
instead, we're going to have to be more selective about which operations
we benchmark for the time being. To support doing that at scale, this
commit adds support for passing *multiple* comma separated lists of query
operations on the command line to the benchmark subcommand, and
permuting all of the lists individually.