Add a verbose flag to the benchmark subcommand
It's useful when debugging queries that cause noria to panic to actually
be able to see those queries beforehand, since noria panicking
unfortunately causes the client to hang rather than return an error.
This adds a --verbose/-v flag to the query_generator benchmark
subcommand to enable printing of each query as we benchmark it.

We may want to have this flag do more things in the future, like print
the DDL as well.