Persist Taster state to redis
Rather than storing Taster's full state in-memory in a hashmap wrapped
in a mutex that can easily be poisoned, persist the benchmark history to
Redis serialized using bincode.

Since the way the taste function uses the state currently is that it
reads the entire set of latest results for a branch all at once but only
saves one benchmark at a time, I've elected to implement this as a redis
map per branch, keyed by the name of the benchmark and with values
containing the serialized format of the map from metric to benchmark
result.