benchmarks: add `forward_time`; measure warm write time; change table
The benchmark suite now measures how long it takes to do a warm write (a write
that propagates fully through the dataflow graph, instead of being dropped due
to a lack of materialization).

This is accomplished by extending the data generator to be able to generate new
unique values given a monotonically increasing counter, which is stored as part
of the `QueryState`. API consumers can now call `make_unique_key()` to generate
a new key for the current query which is guaranteed not to refer to any existing
data; the `generate_data()` function, when called with `make_unique: true`, will
generate the corresponding data to be later inserted.

The benchmark suite uses this to upquery for a new unique key (to make sure the
key is materialized as empty throughout the graph), and then inserts the
generated unique data, timing how long it takes for a blocking read of that key
to complete.

Additionally, the `upquery_forward_time` benchmark metric was renamed to
`upquery_time`, and changed to do something that makes sense; before, the code
didn't actually use the correct metrics for what it wanted to achieve. A new
`forward_time` metric records the time spent forwarding dataflow outside of
upqueries.

(Oh, and the output table is now more concise and includes 100% more cringey
emoji.)