Use a HashMap instead of a Map for base table timestamps.
The current map API assumes that the keys in the map are
contiguous and zero-indexed, for example if a LocalNodeIndex
of 4 is inserted into the map, keys() will return 0..4.
This doesn't match how timestamps are currently used for example:
(1) timestamps may only includei the set of base tables included
in all upstream data flow nodes, and (2) timestamps sent from
debezium to a base table only include that base tables timestamp.

This change will trade off map access performance benefits
for a lower memory and network footprint for the timestamp object
and packet.