Allow determining external address from AWS
When noria is running in distributed mode in any way - even if it's just
a noria-mysql instance running on a different machine - the noria
instances themselves needs to advertise the IP they're available at to
the other instances via zookeeper. Previously, we were mixing the *bind*
address for the workers with that publicly-accessible IP address,
meaning that servers that bound to 0.0.0.0 would advertise themselves as
accessible at eg 0.0.0.0:48123, which obviously didn't work from the
perspective of other instances. This change adds two new top-level
command-line flags - an --external-address flag, which specifies the
address that the currently running noria instance is accessible at to
other instances, and a --use-aws-external-address flag, which indicates
that we should query the AWS metadata service to determine the private
IP address of the instance and use *that* as the external-adddress
instead.

Everywhere we now bind to a TCP port, we can now use the external
address, whether directly configured or obtained from the AWS metadataa
service, to replace the IP address in the bound address.