Handle NVME block devices when mounting volumes
Sometimes - and for reasons I don't really understand - AWS exposes EBS
volumes as *numbered* nvme block devices instead of the block device
names they're *explicitly configured with*. If this happens, the only
way to figure out which block device is which from the perspective of
the block device names returned from the AWS metadata service is to use
the nvme CLI to print out the vendor-specific extension data as *binary*
and index into it at a fixed index. This is absolutely bonkers, but I
have something that uses the nvme-cli and a cursed pile of shell
pipelining to successfully search for the name of the nvme block device
corresponding to the block device given to us by the AWS metadata
service.

Baking this into the user-data script is getting a *little* out of hand
- we likely want to change this to just be a script inside of the AMI at
some point - but this works for now.