Use templatefile() instead of data.template_file
I'm not sure what it is - potentially a terraform bug - but something
about passing in known-after-apply data to a template_file *data source*
causes it to be marked as requiring replacement every apply, meaning
that if we use it as the user_data of an aws_instance that aws instance
requires a recreation on every apply - this was why we had
lifecycle.ignore_changes set to user_data on both the noria_server and
noria_mysql instances. It *appears*, however, that this bug doesn't
apply to the templatefile *function* - meaning that we can get rid of
the ignore_changes if we replace the data source references for
user_data with calls to that function.