Don't run a noria server as part of dbz connector
For some reason, the debezium connector was running an *entire noria
server instance* in a background thread, rather than just running as a
client. This swaps that out to just use a client connection in the same
way that noria-mysql does.

The two binaries, query-installer and view-checker, that were originally
part of the debezium-connector and also started a noria server rather
than connecting as a client, have been moved to a new noria/tools crate,
and updated to also connect as a client instead of running a server.