Fix all clippy lints in debezium connector
We're getting ready to deploy the debezium connector to production, and
many clippy lints give good hints about performance and stability, so
this does a quick pass on fixing all present clippy lints in the crate.
The bulk of the change here is actually just refactoring the constructor
for the DebeziumConnector itself into a Builder, which was triggered by
clippy's max-arguments lint on the function, but is a good idea
generally  because it's *very* easy to get argument order mixed up (I
was already starting to do it!). Besides that there's a couple of ok_or
-> ok_or_else and one case where we could replace a string with a char.