Implement Node.js client
Implement a javascript client for readyset for use in Node.js
applications.

js-client is set up as a node package, with the idea that it
will be published to npm.

The high-level design is a rust JsClient struct that is wrapped by a
special set of functions that utilize Neon bindings.
These Neon bindings allow a rust function to take in arguments with
javascript types and return values with javascript types. Neon then
allows these functions to be exported as a native Node.js addon and
imported into javascript like any other Node.js package. These
functions are then wrapped by a javascript Client class, which provides
the actual client library designed to be used by customers.

js-client's rust backend is located in js-client/src/lib.rs.
This file contains a JsClient struct that wraps the
Backend struct from noria-client (which is the core of all clients
connecting to Noria). This file then utilizes the Neon package
(the napi version) to create and export 4 functions:
connect, async_prepare, async_execute*, and async_query.
The connect function returns a rust struct inside of a Neon JsBox, which
is essentially like returning a pointer to the struct that the
javascript garbage collector knows to clean up after. The async_prepare,
async_execute*, and async_query functions spawn new threads to do their heavy
lifting and then execute javascript callbacks to return their results to
Node.js once they are done.

js-client's javascript frontent is located in js-client/lib/index.js.
This file contains a Client class. When a new Client object is
created, it creates a new JsClient struct and uses it to perform its
queries. The Client's prepare, execute*, and query functions are async
and utilize the Promise framework. The Client constructor takes in a
config object which has a number of different options to be configured
when creating a backend struct. Any of these options can be omitted to
use the default value instead.

An example of querying with the javascript library is located in
js-client/examples.

*NOTE: async_execute/execute is not yet implmented because we are
waiting for changes to ParamParser to be finished before implementing.