Connect to the internal IP of the docs instance
Use the VPC-internal IP of the docs instance, which is accessible via
VPC peering from the VPC that the buildkite instances are running in, to
copy up docs, rather than the external IP resolved via the domain name.

This is necessary because the docs instance no longer has port 22
globally open - it now only allows connection from inside the VPC (with
specifically the VPC CIDR block for the buildkite VPC allowed).