Fixed projection bug. Ticket ID 264.
Before this, when a column was aliased to the name
of a column used in a previous node, it was taking that
column's results instead of the correct results. This is
because, when determining the source of the column on the parent node,
the code was using column_1 == column_2. The equals implementation on
this struct is majorly effed up (documented in ticket 326)

This is somewhat a bandaid fix, as the new method merely checks the
alias and then checks equality rather than jumping straight to equality
but it seemed like the best short term solution.