Fixed sharding bug. Ticket ID 263.
Specifically, when the source was sharded by the same column index
as the key column in the reader node, incorrect shards were queried.
This was an incorrect optimization made to try to avoid querying
unnecessary shards. The correct optimization is to only check the shards
where the SOURCE of the key columns are.

We already had the source in scope by calling keys#provenance_of() so
the fix was just a matter of correctly using those values.

In addition, I split the failing test case into 2 separate cases, one
with a parameter and one without in order to preserve test case
simplicity.