Make debezium connector recover from errors
Previously, any error encountered during handling of a message in the
debezium connector would propagate up the error handling chain, causing
the whole program to crash. This is basically never what we want - if we
encounter a (non-panic) error while handling a message, we'd like to
keep handling messages afterwards!

There is a decision we have to make about errors, though: some errors
are intrinsic to the message, and will be encountered every time we
attempt to handle that message - these are things like the message body
not being valid utf8 or json decoding failing. On the other hand, some
errors are entirely our fault, and the message that caused them can
be retried in the future - these are things like being unable to connect
to noria. To aid in handling these different types of errors, this
commit defines a new MessageError enum that differentiates between those
two cases (Invalid and Internal, respectively), and when handling a
message we commit the handling to kafka for all invalid messages, and
don't for all internal ones (so that they can be retried later).