Add debezium connector support to terraform module
Add support to the noria terraform module for running debezium, kafka,
and the debezium connector all connected to a (externally configured and
passed-in) RDS database on one end and Noria on the other end. Similarly
to both noria-server and zookeeper, none of this is running in an HA
configuration, but everything is running on its own instance so there's
room to scale up in the future if we need to make that possible.

Since it seems like it's basically the only supported way to run it, the
debezium connector is running in a docker container (though it's still
created and managed by a systemd service). Also, since it appears
there's no way in debezium to statically configure data sources, the
user-data init script for this instance has to send an HTTP request via
curl to the connector to set up the initial data source for the RDS DB.
Since the actual *data* here is stored in zookeeper, meaning the
lifecycles don't match up properly, this is quite brittle - in the
future I think probably the best thing to do here is to bake the
management of the data sources for debezium into the debezium connector
itself - that way we can check to see if the connection already exists,
reconfigure it if we need to, etc. But for now, this is what we've got.