Reset systemd failure count in init scripts
Systemd maintains a time-based counter per service of how frequently
restart requests have been queued by that service - but doesn't
discriminate between restart requests due to failure and explicitly
initiated (by invoking `systemctl restart`) restart requests. This means
that when we're setting up a new instance, and services are in a restart
loop due to not being configured yet, and we run systemctl restart after
configuring the instance, that restart requests counts towards the
limit, causing it to fail. This commit adds a call to `systemctl
reset-failed` before the call to `systemctl restart` in all the
provisioning scripts, which resets the failure counter so that the
subsequent restart call actually succeeds. Also, *before* we start
configuring the service we stop it, which doesn't seem to be strictly
necessary but does prevent the service from queuing up a bunch of
restart requests while we're trying to configure it, so it can't hurt.