Update nom-sql to the 2018 edition
Mark nom-sql as edition = "2018" and update all the relevant pieces to
get it compiling again. The future is now!

Basically all of this is just adding `crate::` before various `use`
statements for crate-local modules - I ostensibly *could have* removed
some `#[macro_use] extern crate`s but I elected not to in order to keep
the change as minimal as possible.