Test a simplified version of Weavegrid's schema
Weavegrid sent us a subset of their database schema that they're going
to be running test queries on. This adds a pared down version of that
schema (with some things like ON UPDATE clauses and various KEYs in
tables removed) as a logictest, both to make sure the tables parse
successfully and as a starting point for being able to actually test the
queries they want to run in the future.