Fix all clippy lints in nom-sql
I'm going to be making a lot of changes in nom-sql soon, so now seems
like as good a time as any to auto-fix a bunch of clippy lints and
enable clippy linting in buildkite.yaml. The vast majority of these
changes were just running `cargo clippy --fix -Z unstable-options` and
committing the result. There were only two places where I had to
explicitly *allow* the clippy lints - one complex_type that I only
really elected not to refactor to keep the diff smaller, and one
large_enum_variant warning that we *should* benchmark at some point, but
I've elected not to here again to keep the diff smaller.