Allow users to replicate readers into a given worker.
There's a new endpoint in the controller to handle reader replication requests,
where users can request that, given a query or set of queries, a reader node for
 each of them is replicated.
The user may also specify the worker where the reader should be replicated into.
Even though these reader nodes are replicas from a user point of view, there's
no such distinction in Noria: since readers are leaf nodes (therefore, can't
have children), then there's no risk of messing replay paths and simply
"cloning" the nodes is enough.
Regarding worker assignment, although it's impossible to assign nodes to
workers, we know that reader nodes belong to their own thread domain, which can
be assign to workers, so in the end the new domains go to the user-specified
worker (if provided, otherwise the traditional domain-worker assignment
mechanism is used).