Add the start of an Expression enum to nom-sql
Going forward, I'd like to do some pretty substantial refactoring to the
AST to nom-sql to simplify it significantly - there's a lot involved
there (some of it is talked about in
https://app.clubhouse.io/readysettech/story/101) but the core of it is
getting rid of a lot of the ad-hoc AST nodes (I'm looking at you,
FunctionArgument, ColumnOrLiteral, ArithmeticBase, etc.) and replacing
them with just a single Expression type, which recursively includes
itself as a member in various positions. This commit adds the very
beginning of that Expression type - without actually *using* it in the
AST anywhere - so we at least have a starting point that we can start to
refactor things to use.