Support CAST expressions in projections
Add support for CAST expressions all the way from parsing them through
evaluating them as part of ProjectExpression::eval, by calling
DataType::coerce_to.

Since this is the first instance of an expression (not just a column or
literala) supported by Project that isn't just "arithmetic", this also
involves a reasonably substantial rename and refactor to a lot of the
way the data flows through to project. A couple notable things here:

- Previously, any instance of a "function" on a column (ugh) was
  considered to be an aggregate - because all of the functions *were*
  aggregates! This had to be changed so that non-aggregate function
  expressions flow through to the project node creation rather than
  the grouped node creation
- I've taken this opportunity to *start* converting some of the path
  from nom-sql to MIR to use the new Expression enum instead of all the
  ad-hoc AST types - nothing in nom-sql uses this yet, but when storing
  expressions for projected columns in MIR we convert to it. This should
  make the eventual "make the expression AST make sense" refactor a lot
  easier
- Lots of places we previously used the word "arithmetic" now use the
  word "expression"